pyproject.toml
[tool.poetry]
name = "spa"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^6.0"
djangorestframework = "^3.16.1"
pillow = "^12.0.0"
"django-filter" = "^24.3"
"python-dotenv" = "^1.0.1"
psycopg2-binary = "^2.9.11"
djangorestframework-simplejwt = "^5.5.1"
drf-spectacular = "^0.27.2"
stripe = "^9.0.0"
celery = "^5.6.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
